requirements.txt
altair==5.5.0
asn1crypto==1.5.1
attrs==25.4.0
blinker==1.9.0
boto3==1.40.72
botocore==1.40.72
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
confluent-kafka==2.12.2
cryptography==46.0.3
delta-spark==4.0.0
duckdb==1.4.2
filelock==3.20.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
importlib-metadata==6.11.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.11.0
numpy==1.26.4
packaging==23.2
pandas==2.1.4
pillow==10.4.0
platformdirs==4.5.0
plotly==5.18.0
protobuf==4.25.8
psycopg2-binary==2.9.11
py4j==0.10.9.9
pyarrow==22.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyspark==4.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.28.0
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.0.0
sortedcontainers==2.4.0
streamlit==1.29.0
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
wheel==0.45.1
zipp==3.23.0
